Rewrite & optimize the finding algorithm
The state of current iteration of the algorithm is stored inside a single map.
This makes the algorithm more concise, as we don't need to pass three
separate arguments to each function.

Also, we made two important performance optimizations here:

1. Indexes to visit are stored in a set instead of a vector. This makes
sure we won't visit the same index twice.

2. When extracting the current index from the indexes to visit,
we called `rest` on that coll. This converted the coll to a list
and we would loose the semantics of a set. Now we make sure
to convert it back to a set.